Add feature to resize video
VideoPlayer.js now uses React Hooks instead of
ES6 classes.
useWindowSize.js is created to get the current browser size.
Using this hook, VideoPlayer get width and height of current browser, and resizes player if they are too small
to display 720*480 videos.
The overall usage is the same as before.