Finalize cookie setting
The video state checker is now seperated from the main page as a
component.
The component now has a concrete layout using material-ui
instead of a rough layout with texts as before.
The lecture page now collects only the lastLectureId and videoEnd state,
and the component is responsible for expressing it to the user.